# api.yaml
openapi: 3.0.3
info:
  title: Gestión de seguimiento docente
  description: |-
    API para gestionar el seguimiento docente
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.1.0
servers:
  - url: http://localhost:8070/api/v1/
...
paths:
...
  /login:
    post:
      tags:
        - usuarios
      summary: Devuelve los parámetros de un usuario
      description: Dados un email y una contraseña válidas y en base de datos, devuelve los parámetros del usuario asociado.
      operationId: authenticateUser
      requestBody:
        $ref: '#/components/requestBodies/LoginUsuario'
      responses:
        '200':
          $ref: '#/components/responses/UsuarioData'
        '422':
          description: Datos no válidos
  /usuarios:
    get:
      tags:
        - usuarios
      summary: Devuelve una lista de usuarios
      description: Devuelve la lista de los usuarios de la base de datos
      operationId: getUsuarios
      responses:
        '200':
          $ref: '#/components/responses/UsuarioListaData'

    post:
      tags:
        - usuarios
      summary: Crea un usuario en la base de datos
      description: Dados los datos de un usuario, y el usuario que intenta crear a dicho usuario, crea a un usuario en la base de datos.
      operationId: createUser
      requestBody:
        $ref: '#/components/requestBodies/CreateUsuario'
      responses:
        '201':
          description: Usuario creado con éxito
        '404':
          description: Usuario creador no válido
        '422':
          description: Datos no válidos
  /usuarios/{idUsuario}:
    get:
      tags:
        - usuarios
      summary: Devuelve los datos de un usuario por su id
      description: Devuelve los datos del usuario con id = {idUsuario}
      operationId: getUsuarioById
      parameters:
        - $ref: '#/components/parameters/idUsuario'
      responses:
        '200':
          $ref: '#/components/responses/Usuario'
        '400':
          description: Id suministrado no válido
        '404':
          description: Usuario no encontrado
...
  /excepciones:
    post:
      tags:
        - excepciones
      summary: Crea una nueva excepción en la base de datos
      description: Dados los parámetros, genera una excepción para una actividad en la base de datos
      operationId: createExcepcion
      requestBody:
        $ref: '#/components/requestBodies/CreateExcepcion'
      responses:
        '200':
          description: Excepción creada con éxito
        '400':
          description: Id suministrado no válido
        '404':
          description: Actividad no encontrada
        '422':
          description: Datos no válidos
  /excepciones/{idExcepcion}:
    get:
      tags:
        - excepciones
      summary: Devuelve información de una excepción por su id
      description: Devuelve información sobre la excepción con id = {idExcepcion}
      operationId: getExcepcionById
      parameters:
        - $ref: '#/components/parameters/idExcepcion'
      responses:
        '200':
          $ref: '#/components/responses/ExcepcionData'
        '400':
          description: Id suministrado no válido
        '404':
          description: Excepcion no encontrada
  /excepciones/actividades/{idActividad}:
    get:
      tags:
        - excepciones
        - actividades
      summary: Devuelve una lista de excepciones de una actividad por su id
      description: Devuelve la lista de excepciones asociadas a la actividad con id = {idActividad}
      operationId: getExcepcionesOfActividad
      parameters:
        - $ref: '#/components/parameters/idActividad'
      responses:
        '200':
          $ref: '#/components/responses/ExcepcionListaData'
        '400':
          description: Id suministrado no válido
        '404':
          description: Actividad no encontrada
...
components:
  parameters:
...
    idUsuario:
      name: idUsuario
      in: path
      description: ID del usuario
      required: true
      schema:
        type: integer
        format: int64
      example: 1
    idActividad:
      name: idActividad
      in: path
      description: ID de la actividad
      required: true
      schema:
        type: integer
        format: int64
      example: 1
...
    idExcepcion:
      name: idExcepcion
      in: path
      description: ID de la excepcion
      required: true
      schema:
        type: integer
        format: int64
      example: 1
...
  schemas:
    conId:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          description: Id de la entidad
      required:
        - id
    conMod:
      type: object
      properties:
        creadoEn:
          type: string
          format: date-time
          description: Fecha y hora de creación de la entidad
        actualizadoEn:
          type: string
          format: date-time
          description: Fecha y hora de la última actualización de la entidad
        creadoPor:
          type: integer
          format: int64
          example: 1
          description: Id del usuario que ha creado la entidad
        actualizadoPor:
          type: integer
          format: int64
          example: 2
          description: Id del último usuario que ha actualizado la entidad
      required:
        - creadoEn
        - actualizadoEn
        - creadoPor
        - actualizadoPor
...
    UsuarioBasicData:
      type: object
      properties:
        email: 
          type: string
          description: Email del usuario
          example: 'grhopper@gmail.com'
    UsuarioLoginData:
      allOf:
        - $ref: '#/components/schemas/UsuarioBasicData'
        - type: object
          properties:
            password:
              type: string
              description: Contraseña del usuario
              example: "As34dnF5fsDlRPfdnsmRSDDFJAJ4JDRJEJaSD"
    UsuarioData:
      allOf:
        - $ref: '#/components/schemas/UsuarioBasicData'
        - type: object
          properties:
            nombre:
              type: string
              description: Nombre del Usuario
              example: Grace
            apellidos:
              type: string
              description: Apellidos del Usuario
              example: Hopper
            rol:
              type: string
              description: Rol del Usuario
              example: Admin
    UsuarioCreateData:
      allOf:
        - $ref: '#/components/schemas/UsuarioData'
        - $ref: '#/components/schemas/UsuarioLoginData'
        - type: object
          properties:
            creador:
              type: integer
              format: int64
              description: Id del usuario
              example: 67
    Usuario:
      allOf:
        - $ref: '#/components/schemas/UsuarioData'
        - $ref: '#/components/schemas/conId'
        - $ref: '#/components/schemas/conMod'
    ListaUsuarios:
      type: object
      properties:
        usuarios:
          type: array
          items: 
            $ref: '#/components/schemas/conId'
...
    ExcepcionData:
      type: object
      properties:
        esta_reprogramado:
          type: string
          enum:
            - Sí
            - No
          description: Si está o no reprogramada la actividad
        esta_cancelado:
          type: string
          enum:
            - Sí
            - No
          description: Si está o no cancelada la actividad (aunque esté reprogramada, se puede cancelar)
        actividad_id:
          type: integer
          format: int64
          description: Id de la actividad a la que se asocia la excepción
          example: 1
        fecha_inicio_act:
          type: string
          description: fecha de inicio de la actividad (o instancia de actividad) cancelada/reprogramada
          example: '2022-03-01 10:00:00.000'
        fecha_fin_act:
          type: string
          description: fecha de fin de la actividad (o instancia de actividad) cancelada/reprogramada
          example: '2022-03-01 11:40:00.000'
        fecha_inicio_ex:
          type: string
          description: fecha de inicio de la reprogramación de la actividad (o instancia de actividad) reprogramada
          example: '2022-03-05 13:00:00.000'
        fecha_fin_ex:
          type: string
          description: fecha de fin de la reprogramación de la actividad (o instancia de actividad) reprogramada
          example: '2022-03-05 14:40:00.000'
    ListaExcepciones:
      type: object
      properties:
        excepciones:
          type: array
          items:
            $ref: '#/components/schemas/conId'
          description: Array con ids de excepciones válidas
          example: [ { "id": 1 } ]
...
  requestBodies:
...
    LoginUsuario:
      required: true
      description: Email y Contraseña de un usuario existente
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UsuarioLoginData'
    CreateUsuario:
      required: true
      description: Datos necesarios para la creación de un usuario
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UsuarioCreateData'
...
    CreateExcepcion:
      required: true
      description: Datos necesarios para la creación de una excepción
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ExcepcionData'
...
  responses:
...
    UsuarioData:
      description: Datos de un usuario a devolver en un login
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UsuarioData'
    Usuario:
      description: Datos de un usuario mas su id
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Usuario'
    UsuarioListaData:
      description: Lista de ids de usuarios a devolver a petición de una query
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ListaUsuarios'
...
    ExcepcionData:
      description: Datos de una excepcion
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ExcepcionData'
    ExcepcionListaData:
      description: Lista de ids de excepciones a devolver a petición de una query
      content:
        application/json:
          schema:
             $ref: '#/components/schemas/ListaExcepciones'
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY